# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "LTI Canvas Devops Backend",
    "description": "LTI Canvas Devops API Documentation"
  },
  "schemes": [
    "http",
    "https"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
...
  "paths": {
    "/settings/get_tool_categories": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get tool categories by application",
        "description": "Returns tool categories based on application key",
        "parameters": [
          {
            "name": "application_key",
            "in": "query",
            "description": "Application Key of the application",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Tool Categories",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "category_name": {
                    "type": "string",
                    "example": "planning"
                  },
                  "instance_count": {
                    "type": "number",
                    "example": 1
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          }
        }
      }
    },
    "/settings/get_tools": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get tools by application and tool category",
        "description": "Returns tools based on application key and tool category",
        "parameters": [
          {
            "name": "application_key",
            "in": "query",
            "description": "Application Key of the application",
            "type": "string",
            "required": true
          },
          {
            "name": "tool_category",
            "in": "query",
            "description": "Tool Category",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Tools",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "tool_url": {
                    "type": "string",
                    "example": "http://3.87.19.19:8081"
                  },
                  "tool_instance_name": {
                    "type": "string",
                    "example": "Public Jira"
                  },
                  "tool_name": {
                    "type": "string",
                    "example": "JIRA"
                  },
                  "tool_version": {
                    "type": "string",
                    "example": "v1"
                  },
                  "proxy_required": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          }
        }
      }
    },
    "/settings/sync_tool_projects": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Synchronize tool projects",
        "description": "Returns response based on synchronization status",
        "parameters": [
          {
            "name": "application_key",
            "in": "query",
            "description": "Application Key of the application",
            "type": "string",
            "required": true
          },
          {
            "name": "tool_category",
            "in": "query",
            "description": "Tool Category",
            "type": "string",
            "required": true
          },
          {
            "name": "tool_instance_name",
            "in": "query",
            "description": "Tool Name",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Synchronization is successfull",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Projects synchronization is successful"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          }
        }
      }
    },
    "/settings/sync_tool_users": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Synchronize tool users",
        "description": "Returns response based on synchronization status",
        "parameters": [
          {
            "name": "application_key",
            "in": "query",
            "description": "Application Key of the application",
            "type": "string",
            "required": true
          },
          {
            "name": "tool_category",
            "in": "query",
            "description": "Tool Category",
            "type": "string",
            "required": true
          },
          {
            "name": "tool_instance_name",
            "in": "query",
            "description": "Tool Name",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Synchronization is successfull",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Projects synchronization is successful"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/settings/validate_tool": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Validate a new tool",
        "description": "Validates a new tool based on the input data",
        "parameters": [
          {
            "name": "tool_details",
            "in": "body",
            "description": "Tool Input Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "application_key": {
                  "type": "string",
                  "example": "app1"
                },
                "tool_url": {
                  "type": "string",
                  "example": "http://3.87.19.19:8081"
                },
                "tool_instance_name": {
                  "type": "string",
                  "example": "Public Jira"
                },
                "tool_category": {
                  "type": "string",
                  "example": "Planning"
                },
                "tool_name": {
                  "type": "string",
                  "example": "JIRA"
                },
                "tool_version": {
                  "type": "string",
                  "example": "v1"
                },
                "proxy_required": {
                  "type": "boolean",
                  "example": false
                },
                "tool_auth": {
                  "type": "object",
                  "properties": {
                    "auth_type": {
                      "type": "string",
                      "example": "Basic"
                    },
                    "auth_username": {
                      "type": "string",
                      "example": "admin"
                    },
                    "auth_password": {
                      "type": "string",
                      "example": "admin"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tool Validation is successfull",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Tool Validation is successfull"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          }
        }
      }
    },
    "/settings/save_tool": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Validate a new tool",
        "description": "Validates a new tool based on the input data",
        "parameters": [
          {
            "name": "tool_details",
            "in": "body",
            "description": "Tool Input Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "application_key": {
                  "type": "string",
                  "example": "app1"
                },
                "tool_url": {
                  "type": "string",
                  "example": "http://3.87.19.19:8081"
                },
                "tool_instance_name": {
                  "type": "string",
                  "example": "Public Jira"
                },
                "tool_category": {
                  "type": "string",
                  "example": "Planning"
                },
                "tool_name": {
                  "type": "string",
                  "example": "JIRA"
                },
                "tool_version": {
                  "type": "string",
                  "example": "v1"
                },
                "proxy_required": {
                  "type": "boolean",
                  "example": false
                },
                "tool_auth": {
                  "type": "object",
                  "properties": {
                    "auth_type": {
                      "type": "string",
                      "example": "Basic"
                    },
                    "auth_username": {
                      "type": "string",
                      "example": "admin"
                    },
                    "auth_password": {
                      "type": "string",
                      "example": "dmin"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tool Validation is successfull",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Tool Validation is successfull"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          }
        }
      }
    },
...
  },
...
}